Add run manager step
This commit adds an additional step for a run manager that can determine
if individual runs need to be started/continued and submit them as
separate batch jobs.  The concept works, but compass requires `compass
run` commands to be issued from compute nodes, so this step can only be
run from an interactive node or batch job itself, which is otherwise
unnecessary.  It might be better to have the run manager be a separate
script called from outside the compass system.