Have setup store the name of the machine
This commit replaces changes in e874630b0eb438ba94b9c899aa0d4f4c23a73e91
by having the setup command store the name of the machine in the config.
This avoids needing each machine file to have a option saying the name
of the machine.  This commit also undoes the change to the
cori-haswell.cfg file in the previous commit.  This cleaner solution was
suggested by Xylar in review.