Remove unused argument for making mapping files
The 64-bit offset argument was missing a dash.  But that was ok, because
it is incompatible with the netcdf4 argument.  Because it was missing
dash, it was not being used.  This commit removes the argument fully to
avoid confusion, but this does not affect how esmf_regridweightgen was
actually running.